Refactor pytest_beets_plugin
